refactor: 폴더구조 정리 및 출석 리팩토링
