+++ src/app/meetings/[meetingId]/hooks/useMeetingDetailQueries.ts
@@ -1,0 +1,36 @@
+"use client";
+
+import { useQuery } from "@tanstack/react-query";
+import {
+  fetchMeetingDetail,
+  fetchMeetingParticipants,
+  fetchMeetingRecommendationCandidates,
+} from "@/app/meetings/[meetingId]/api/meeting-detail.api";
+import {
+  getMeetingDetailQueryKey,
+  getMeetingParticipantsQueryKey,
+  getMeetingRecommendationCandidatesQueryKey,
+} from "@/app/meetings/[meetingId]/model/meeting-detail.query-keys";
+
+export function useMeetingDetailQueries(meetingId: number) {
+  const detailQuery = useQuery({
+    queryKey: getMeetingDetailQueryKey(meetingId),
+    queryFn: () => fetchMeetingDetail(meetingId),
+  });
+
+  const participantsQuery = useQuery({
+    queryKey: getMeetingParticipantsQueryKey(meetingId),
+    queryFn: () => fetchMeetingParticipants(meetingId),
+  });
+
+  const recommendationCandidatesQuery = useQuery({
+    queryKey: getMeetingRecommendationCandidatesQueryKey(meetingId),
+    queryFn: () => fetchMeetingRecommendationCandidates(),
+  });
+
+  return {
+    detailQuery,
+    participantsQuery,
+    recommendationCandidatesQuery,
+  };
+}
