+++ src/app/meetings/[meetingId]/model/meeting-detail.query-keys.ts
@@ -1,0 +1,8 @@
+export const getMeetingDetailQueryKey = (meetingId: number) =>
+  ["meeting-detail", meetingId] as const;
+
+export const getMeetingParticipantsQueryKey = (meetingId: number) =>
+  ["meeting-participants", meetingId] as const;
+
+export const getMeetingRecommendationCandidatesQueryKey = (meetingId: number) =>
+  ["meeting-recommendation-candidates", meetingId] as const;
