+++ src/app/meetings/modal/model/meeting-image.types.ts
@@ -1,0 +1,4 @@
+export interface UploadImageResponse {
+  presignedUrl: string;
+  publicUrl: string;
+}
